refactor: article관련 리팩토링
