+++ lib/utils/formatDate.ts
@@ -1,4 +1,4 @@
-export function formatDate(dateString: string) {
+export function formatDate(dateString: Date) {
   const date = new Date(dateString);
   const year = date.getFullYear();
   const month = String(date.getMonth() + 1).padStart(2, "0"); // 월은 0부터 시작하므로 1을 더함
